+++ frontend/src/pages/host/mainPage/HostMainPage.tsx
@@ -1,11 +1,11 @@
 import { IoLogoInstagram } from 'react-icons/io5';
 import { MdEmail } from 'react-icons/md';
-import FooterLogo from '../../../@assets/footer-logo.svg?react';
+import FooterLogo from '../../../@assets/logo/footer-logo.svg?react';
 import Button from '../../../components/@common/buttons/button/Button';
 import IconButton from '../../../components/@common/buttons/iconButton/IconButton';
 import { DividerLine } from '../../../styles/@common/DividerLine.styles';
 import { createInstagramUrl } from '../../../utils/createExternalLinks';
-import * as MainPageStyles from '../../mainPage.common.styles';
+import * as MainPageStyles from '../../MainPage.common.styles';
 import { mockData } from '../../mockData';
 
 const HostMainPage = () => {
